Rename route_contexted_reducer_helper -> namespaced_state_reducer_helper.
