+++ tensorboard/webapp/app_routing/BUILD
@@ -161,9 +161,9 @@ tf_ng_module(
 )
 
 tf_ts_library(
-    name = "route_contexted_reducer_helper",
+    name = "namespaced_state_reducer_helper",
     srcs = [
-        "route_contexted_reducer_helper.ts",
+        "namespaced_state_reducer_helper.ts",
     ],
     deps = [
         ":internal_utils",
@@ -210,7 +210,7 @@ tf_ng_module(
         "internal_utils_test.ts",
         "location_test.ts",
         "programmatical_navigation_module_test.ts",
-        "route_contexted_reducer_helper_test.ts",
+        "namespaced_state_reducer_helper_test.ts",
         "route_registry_module_test.ts",
         "store_only_utils_test.ts",
     ],
@@ -220,7 +220,7 @@ tf_ng_module(
         ":internal_utils",
         ":location",
         ":programmatical_navigation_module",
-        ":route_contexted_reducer_helper",
+        ":namespaced_state_reducer_helper",
         ":route_registry",
         ":store_only_utils",
         ":testing",
