+++ tensorboard/webapp/app_routing/namespaced_state_reducer_helper.ts
