+++ tensorboard/webapp/core/store/BUILD
@@ -12,7 +12,7 @@ tf_ng_module(
         "index.ts",
     ],
     deps = [
-        "//tensorboard/webapp/app_routing:route_contexted_reducer_helper",
+        "//tensorboard/webapp/app_routing:namespaced_state_reducer_helper",
         "//tensorboard/webapp/core:types",
         "//tensorboard/webapp/core/actions",
         "//tensorboard/webapp/deeplink",
